refactor!: .mjs 拡張子を廃止し .js に統一する
Node.js v22 で ESM から CJS の import が正式サポートされ、リポジトリの
engines.node は 24 に固定されているため、.mjs を維持する理由がなくなった。
全 package.json は既に "type": "module" を宣言済みなので、拡張子を .js に
リネームしても解釈は変わらない。

BREAKING CHANGE: 公開パッケージの main / exports / bin のパスが
./*.mjs から ./*.js に変わる。
