+++ packages/create-eleventy/eslint.config.js
