+++ packages/create-eleventy/templates/default/eleventy.config.js
