+++ packages/create-eleventy/templates/default/prettier.config.js
