+++ packages/create-eleventy/templates/default/src/__data/site.js
