+++ packages/eleventy-plugin-postcss/package.json
@@ -19,14 +19,14 @@
   "exports": {
     ".": {
       "types": "./index.d.ts",
-      "default": "./index.mjs"
+      "default": "./index.js"
     }
   },
-  "main": "./index.mjs",
+  "main": "./index.js",
   "files": [
     "README.md",
     "index.d.ts",
-    "index.mjs"
+    "index.js"
   ],
   "scripts": {
     "lint": "eslint --fix ."
