+++ packages/eleventy-plugin-preact-island/resolver.js
