+++ packages/eleventy-plugin-preact/package.json
@@ -21,26 +21,26 @@
   "exports": {
     ".": {
       "types": "./index.d.ts",
-      "default": "./index.mjs"
+      "default": "./index.js"
     },
     "./eleventy": {
       "types": "./eleventy.d.ts",
-      "default": "./eleventy.mjs"
+      "default": "./eleventy.js"
     },
     "./loaders/mdx": {
-      "default": "./loaders/mdx.mjs"
+      "default": "./loaders/mdx.js"
     },
     "./loaders/jsx": {
-      "default": "./loaders/jsx.mjs"
+      "default": "./loaders/jsx.js"
     }
   },
-  "main": "./index.mjs",
+  "main": "./index.js",
   "files": [
     "README.md",
     "index.d.ts",
-    "index.mjs",
+    "index.js",
     "eleventy.d.ts",
-    "eleventy.mjs",
+    "eleventy.js",
     "loaders"
   ],
   "scripts": {
